# gateway/api/swagger/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This swagger UI was created to manage personal finance",
        "title": "PERSONAL FINANCE MANAGEMENT",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/api/users/forgot-password": {
            "post": {
                "description": "This method is responsible for sending a password reset code to the user's email",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "USERS"
                ],
                "summary": "Request password reset",
                "parameters": [
                    {
                        "description": "User ID for reset",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/users.UserID"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Code sent to your email",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Error retrieving email",
                        "schema": {}
                    },
                    "403": {
                        "description": "Permission Denied",
                        "schema": {}
                    },
                    "500": {
                        "description": "Error storing reset code or sending email",
                        "schema": {}
                    }
                }
            }
        },
        "/api/users/login": {
            "post": {
                "description": "This method is responsible for user login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "USERS"
                ],
                "summary": "User login",
                "parameters": [
                    {
                        "description": "User login details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/users.UserLogin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful with token",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Error while unmarshaling",
                        "schema": {}
                    },
                    "401": {
                        "description": "Invalid username or password",
                        "schema": {}
                    },
                    "403": {
                        "description": "Permission Denied",
                        "schema": {}
                    }
                }
            }
        },
        "/api/users/logout": {
            "post": {
                "description": "This method is responsible for logging out the user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "USERS"
                ],
                "summary": "User logout",
                "parameters": [
                    {
                        "description": "User ID for logout",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/users.UserID"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Logout successful",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Error while Logginf out user ",
                        "schema": {}
                    },
                    "403": {
                        "description": "Permission Denied",
                        "schema": {}
                    }
                }
            }
        },
        "/api/users/register": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "This method is responsible for registering new users",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "USERS"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "User registration details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/users.UserInfo"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User registered successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "integer"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request body or error unmarshaling",
                        "schema": {}
                    },
                    "403": {
                        "description": "Permission Denied",
                        "schema": {}
                    },
                    "500": {
                        "description": "Error storing user data or sending confirmation email",
                        "schema": {}
                    }
                }
            }
        },
        "/api/users/update-password": {
            "post": {
                "description": "This method is responsible for updating the user's password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "USERS"
                ],
                "summary": "Update user password",
                "parameters": [
                    {
                        "description": "New password details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.NewPass"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Password updated successfully",
                        "schema": {
                            "$ref": "#/definitions/users.UserInfo"
                        }
                    },
                    "400": {
                        "description": "Invalid input or user verification error",
                        "schema": {}
                    },
                    "403": {
                        "description": "Permission Denied",
                        "schema": {}
                    },
                    "500": {
                        "description": "Error while updating password",
                        "schema": {}
                    }
                }
            }
        },
        "/api/users/verification": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "This method is responsible for verifying the user registration code",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "USERS"
                ],
                "summary": "Verify user registration code",
                "parameters": [
                    {
                        "description": "User code details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UserCode"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User created successfully",
                        "schema": {
                            "$ref": "#/definitions/users.UserInfo"
                        }
                    },
                    "400": {
                        "description": "Error verifying code",
                        "schema": {}
                    },
                    "500": {
                        "description": "Error creating user or sending welcome email",
                        "schema": {}
                    }
                }
            }
        },
        "/api/users/{id}": {
            "get": {
                "description": "This method retrieves a user by their ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "USERS"
                ],
                "summary": "Get user by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User data retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/users.UserInfo"
                        }
                    },
                    "400": {
                        "description": "Error while retrieving user",
                        "schema": {}
                    },
                    "403": {
                        "description": "Permission Denied",
                        "schema": {}
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.NewPass": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "model.UserCode": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                }
            }
        },
...
        "users.UserID": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "users.UserInfo": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "users.UserLogin": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}